requirements.txt
Django
django-ninja
django-ninja-jwt[crypto]
rav